requirements.txt
Flask
flask_sqlalchemy
python-dotenv
llama-index-core
llama-index-llms-huggingface
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
transformers
accelerate
torch